make_search_index -> update_composition_index
